Move Test file locations
