+++ src/__mocks__/gatsby.ts
