+++ src/__tests__/Footer.test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Footer from '.';
+import Footer from '../components/Template/Footer';
 
 describe('Footer', () => {
   it('should render without crashing', () => {
