+++ src/__tests__/Icons.test.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import './index';
-import LangIcon from './LangIcon';
-import SocialIcon from './SocialIcon';
+import '../components/Icons/index';
+import LangIcon from '../components/Icons/LangIcon';
+import SocialIcon from '../components/Icons/SocialIcon';
 
 describe('Lang Icon', () => {
   it('should render without crashing', () => {
